% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fscores_maker.R
\name{Fscores_maker}
\alias{Fscores_maker}
\title{Calculate F_1 scores based on SVM evaluations}
\usage{
Fscores_maker(actual, pred)
}
\arguments{
\item{data}{Data frame containing the actual and predicted values of the property under consideration}

\item{actual}{Character string of the column name of the the actual values of the property under consideration}

\item{pred}{Character string of the column name of the the predicted values of the property under consideration.
Must be same length as \code{actual}.}
}
\description{
A function for calculating F_1 scores based on SVM evaluations
}
\examples{
data(tonDemo)
Fscores_maker(tonDemo, "party_id", "class_token")
}
\seealso{
\code{\link[e1071]{svm}}
         \code{\link[tonR]{Fscores_boot}}

Other fscores: \code{\link{Fscores_boot}}
}

