% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lix.R
\name{lix}
\alias{lix}
\title{LIX readability score}
\usage{
lix(actual, pred)
}
\arguments{
\item{n_words}{Integer specifying number of words in a document}

\item{n_breaks}{Integer specifying number of breaks (period, comma, capital letters)}

\item{n_longwords}{Integer specifying number of words longer than six letters in a document}
}
\description{
A function for calculating the LIX readability score
}
\examples{
# To be done
}
