% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lix_ton.R
\name{lix_ton}
\alias{lix_ton}
\title{LIX for Talk of Norway}
\usage{
lix_ton(ids, ton_path = "../talk-of-norway/", cores = 1)
}
\arguments{
\item{ids}{Character with ToN ids for the speeches to calculate LIX from}

\item{ton_path}{Character specifying the relative path to the talk-of-norway repository folder}

\item{cores}{Integer specifying number of processor cores to run the function on. Requires the \code{\link[parallel]{parallel-package}}}
}
\value{
A tibble with id, number of words, number of long words, number of sentences, and lix scores.
}
\description{
A function for calculating the LIX readability score for the Talk of Norway corpus
}
\details{
This is a function that uses the annotated part of the Talk of Norway corpus
for computing LIX readability scores. It is a special wrapper for the more
general \code{\link[tonR]{lix}} function.
}
\examples{
data("tonDemo")
# Computing LIX for whole data
all <- lix_by_id(tonDemo$id, ton_path = "../talk-of-norway/", cores = detectCores()-2)
summary(all$lix)

# Computing LIX for Labor Party (A)
ap <- lix_by_id(tonDemo$id[which(tonDemo$party_id == "A")], ton_path = "../talk-of-norway/", cores = detectCores()-2)
summary(ap$lix)

}
\seealso{
\code{\link[tonR]{lix}}
}

