% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.conll.R
\name{read.conll}
\alias{read.conll}
\title{Read CoNLL-type data from Talk of Norway}
\usage{
read.conll(file)
}
\arguments{
\item{file}{Character string specifying the path to .tsv-file to read into R}
}
\description{
A function for reading the CoNLL data used in the Talk of Norway project for legislative speeches
in the Norwegian parliament (1998-2016). Operates as a wrapper for \code{\link[utils]{read.csv}}
}
\examples{
# If the 'talk-of-norway' repository is placed in the folder below the tonR-package:
speech <- read.conll("../talk-of-norway/data/annotations/tale100001.tsv")
barplot(table(speech$part_of_speech))
}
\seealso{
\code{\link[utils]{read.csv}}
}

