% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ton_corpus.R
\name{ton_corpus}
\alias{ton_corpus}
\title{Corpus of ToN texts}
\usage{
ton_corpus(tonFolder, ids, cores = NULL)
}
\arguments{
\item{tonFolder}{Character vector specifying either absolute or relative path to the talk-of-norway repository folder}

\item{cores}{Integer specifying number of words longer than six letters in a document}

\item{id}{Character string specifying id of the file to read into R.}
}
\value{
a object of class corpus containing the speeches specified by the \code{ids} value. See \code{\link[corpus]{quanteda}} for details.
}
\description{
A function for reading selected ToN texts to a corpus. Builds on
}
\examples{
# Reading specific speeches into corpus
corpus <- ton_corpus(tonFolder = "../talk-of-norway/", ids = c("tale000005", "tale000006"))

summary(test)
quanteda::docnames(test)

# Using ton data for specifying ids
data("tonDemo")
corpus <- ton_corpus(tonFolder = "../talk-of-norway/", ids = tonDemo$id[1:10])

summary(test)
quanteda::docnames(test)

}
\seealso{
\code{\link[corpus]{quanteda}}
}
