% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ton_stems.R
\name{ton_stems}
\alias{ton_stems}
\title{Stem text from TON corpus}
\usage{
write_ton_stems(ton_folder, id, lower = TRUE, rm_punct = FALSE)
}
\arguments{
\item{ton_folder}{Character vector specifying either absolute or relative path to the talk-of-norway repository folder.}

\item{id}{Character string specifying id of the file to read into R.}

\item{lower}{Logical specifying whether to lowercase tokens. Defaults to \code{FALSE}.}

\item{rm_punct}{Logical specifying whether to remove punctuation. Defaults to \code{FALSE}.}

\item{...}{additional arguments}
}
\description{
A function for writing token stems from the Talk of Norway corpus in a \code{\link[tidytext]{tidytext}} compatible format with the
\code{\link[SnowballC]{wordStem}} algorighm from the \code{\link[SnowballC]{SnowballC}} package.
}
\examples{
data("tonDemo")

# Getting stems
stems <- ton_stems(ton_folder = "../../..", tonDemo$id[1:10])
head(stems)

# Lowercase stems and punctuation removal
stems <- ton_stems(ton_folder = "../../..", tonDemo$id[1:10], lower = TRUE, rm_punct = TRUE)
head(stems)

}
