% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ton_stems.R
\name{write_ton_stems}
\alias{write_ton_stems}
\title{Write stemmed text from TON corpus}
\usage{
write_ton_stems(data, datarow, id_var, text_var)
}
\arguments{
\item{rows}{Vector with row numbers for texts to be stemmed}

\item{id_var}{Vector with id tags in the data}

\item{text_var}{Vector in data with the text to be stemmed}

\item{ton_folder}{Character with relative path to the talk-of-norway repository folder}

\item{data}{Data frame containing the texts to be stemmed}
}
\description{
A function for calculating F_1 scores based on SVM evaluations
}
\examples{
# To be done
}

