% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fscores_boot.R
\name{Fscores_boot}
\alias{Fscores_boot}
\title{Bootstrap F_1 scores based on SVM evaluations}
\usage{
Fscores_boot(data, actual, pred, subset_rows, nsim = 1000, seed = 8949, cores = 1)
}
\arguments{
\item{data}{Data frame containing the actual and predicted values of the property under consideration}

\item{actual}{Character string of the column name of the the actual values of the property under consideration}

\item{pred}{Character string of the column name of the the predicted values of the property under consideration.
Must be same length as \code{actual}.}

\item{nsim}{Integer specifying number of bootstraps to do.}

\item{cores}{Integer specifying number of cores to simulate on. For Linux users only...sorry Windows.}

\item{subset_rows}{A vector containing the rownames for the data to bootstrap. For full data use \code{subset_rows = rownames(data)}}
}
\description{
A function for calculating F_1 scores based on SVM evaluations
}
\examples{
# To be done
}
\seealso{
\code{\link[e1071]{svm}}

Other fscores: \code{\link{Fscores_maker}}
}

