% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.conll.R
\name{read.conll}
\alias{read.conll}
\title{Read CoNLL-type data}
\usage{
read.conll(tonFolder, id, keep = "all", rmStopwords = TRUE, rmLength = 1000)
}
\arguments{
\item{tonFolder}{Character vector specifying either absolute or relative path to the talk-of-norway repository folder}

\item{id}{Character string specifying id of the file to read into R.}

\item{keep_pos}{Character string that specifies what parts-of-speech should be returned. Possible values are:
"subst", "verb", "sbu", "prep", "det", "adj", "clb", "adv", "pron", "<komma>", "konj".}

\item{rmWords}{Character vector of either "no" or of words to remove.}

\item{keep_features}{Character string specifying which features to extract. Defaults to c("", "", "", "", "",)}
}
\value{
A data frame of the ToN speech specified by \code{id}, with sentence and token boundries, tokens, lemmatized
tokens, parts-of-speech tagging, and morpheme features.
}
\description{
A function for reading the CoNLL data used in the Talk of Norway project for legislative speeches
in the Norwegian parliament (1998-2016). Works as a wrapper for \code{\link[utils]{read.csv}}
}
\examples{
# If the 'talk-of-norway' repository is placed in the folder below the tonR-package
speech <- read.conll("../talk-of-norway/", "tale100001")
barplot(table(speech$part_of_speech))

# Loading a subset of speeches using several ToN id tags
data("tonDemo")

texts <- lapply(tonDemo$id[1:10], function(x){
  read.conll(tonFolder = "../talk-of-norway/", id = x, keep_pos = "adj")
})
lapply(texts, head)

}
\seealso{
\code{\link[utils]{read.csv}}
}
