% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.conll.R
\name{read.conll}
\alias{read.conll}
\title{Read CoNLL-type data from Talk of Norway}
\usage{
read.conll(id, keep = "all", rmStopwords = TRUE, rmLength = 1000)
}
\arguments{
\item{tonFolder}{Character vector specifying either absolute or relative path to the talk-of-norway repository folder}

\item{id}{Character string specifying id of the file to read into R.}

\item{keep}{Character string that specifies what parts-of-speech should be returned. Possible values are:
"subst", "verb", "sbu", "prep", "det", "adj", "clb", "adv", "pron", "<komma>", "konj".}

\item{rmWords}{Character vector of either "no" or of words to remove.}
}
\description{
A function for reading the CoNLL data used in the Talk of Norway project for legislative speeches
in the Norwegian parliament (1998-2016). Operates as a wrapper for \code{\link[utils]{read.csv}}
}
\examples{
# If the 'talk-of-norway' repository is placed in the folder below the tonR-package:
speech <- read.conll("../talk-of-norway/data/annotations/tale100001.tsv")
barplot(table(speech$part_of_speech))
}
\seealso{
\code{\link[utils]{read.csv}}
}

